feat: add Result to ProblemDetails.Abstractions
Refactors the ProblemDetails.Abstractions a fair bit, and adds a Result
class to represent the result of a request. This is a way to return
a result from a request that is either a success or a problem details
response.

feat: create ValidationProblemInstance in ProblemDetails.Abstractions.

refactor: move ValidationErrorBuilder to ProblemDetails.Abstractions
  BREAKING-CHANGE: The ValidationErrorBuilder class has been moved to
  the ProblemDetails.Abstractions assembly.

feat: add ProblemExtensionData to ProblemDetails.Abstractions
  BREAKING-CHANGE: The various Extensions properties in
  ProblemDetails.Abstractions have been changed from being a immutable
  array of key-value pairs, to a new ProblemExtensionData type.

refactor: unseal ProblemInstance

chore: add DebuggerDisplay attributes to multiple types